improved the random offset among the legs
We start off offset, then once we trigger we reset the radius so that it’s the same for all legs, but since the we triggered at different times, now our legs our positioned differently, but our radius is the same, so we’ll take the same time to trigger (we wont have one leg being faster and the other slower), but we don’t trigger at the same time.

When stress-testing the movements, sometimes the triggers synchronize, in which case we always track how close all the “accumulated times” for each leg are. If they’re close within a designated threshold, we reinstate the initial incremented radius, so as to “reset” the offset whenever we get into these nuanced situations. So think of this as a fail-safe in case we the triggers synchronize.

Despite all of this, sometimes we get synced offsets, so what we're doing is constantly checking if the accumulated time between each of the feet is nearly the same, and if it is, take that foot and introduce the radius increment to offset it, and we'll reset the radius to be the same as the others. This happens quickly and is a way to autocorrect those uncommon instances where the feet start syncing. Sure, we get a brief period where one foot may drag longer than the rest when autocorrecting, but this is very brief.